# perf: improve processmetadata performance and reduce allocations