Avoid copying byte slices and make ProcessMetadata zero-allocation.
Add a benchmark to reproduce the numbers:

```
name                       old time/op    new time/op    delta
ProcessMetadata/simple-16    7.89µs ± 8%    0.04µs ± 3%   -99.52%  (p=0.000 n=10+10)

name                       old alloc/op   new alloc/op   delta
ProcessMetadata/simple-16    12.6kB ± 0%     0.0kB       -100.00%  (p=0.000 n=10+10)

name                       old allocs/op  new allocs/op  delta
ProcessMetadata/simple-16      5.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
```