@kruskall 2022-08-03
> Since bytes.Cut was added in Go 1.18, we'll need to update the CI pipeline config to build only with Go 1.18. Not sure why we're testing with multiple versions anyway, since this is not a library -- we should just use one version.

Good catch! I updated the job definition. I'll open a new MR to bump the go version (Go 1.19 was also released today)